order_management_system_backend/requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
filelock==3.12.4
future==0.18.3
gevent==23.9.1
greenlet==3.0.0rc3
idna==3.4
pycparser==2.21
PyJWT==1.7.1
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.4
tldextract==3.5.0
tzdata==2023.3
urllib3==2.0.4
zope.event==5.0
zope.interface==6.0
